Get single data frame

Returns a selected data frame from an ev_data container.

Usage

get_ev_data(ev_data, df_name = character(0), df_index = integer(0))

Arguments

ev_data

An ev_data container

df_name

The name of the data frame. Only used if df_index is not supplied

df_index

The numerical index of the data frame within the ev_data container. Overrides df_name if both are supplied.

Value

A data frame

Details

The data frame can be selected by index (1, 2, 3 etc.) or by name (project.table). The names of data frames within an ev_data container can be queried using get_ev_data_names.

An ev_data container can be created using fetch_ev_data.

See also

Examples

if (FALSE) { # \dontrun{

con <- ev_connect()

# read a variable data request file
vars <- read_ev_variables("path/to/variables/file")
# fetch the data from EpiVault
dat <- fetch_ev_data(con, vars)

# get a data frame by name
get_ev_data(dat, df_name = "proj1.tab1")
# get a data frame by index
get_ev_data(dat, df_index = 1)

} # }